refactor : 프로젝트 삭제 모달 컴포넌트 분리 작업 진행
