+++ src/components/common/Button/Button.types.ts
@@ -1,9 +1,11 @@
+import React from 'react';
+
 export interface ButtonProps {
   size: 'small' | 'medium';
-  variant: 'outline' | 'fill' | 'fillGray' | 'text';
+  variant: 'outline' | 'fill' | 'fillGray' | 'text' | 'fillRed';
   hasIcon: boolean;
   isDisabled?: boolean;
   iconPosition?: 'left' | 'right';
-  onClick: () => void;
+  onClick: (e: React.MouseEvent<HTMLButtonElement>) => void;
   text: string;
 }
