--- src/components/modal/DeleteProjectModal.tsx
@@ -1,144 +1,0 @@
-import React, { useState } from 'react';
-
-import InputArea from '@components/common/InputArea';
-import {
-  useDeleteProject,
-  useGetProject,
-} from '@services/project/Project.hooks';
-import styled from 'styled-components';
-
-const ModalHeader = styled.article`
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-  h2 {
-    color: var(--main-black, #000);
-    font-family: Pretendard;
-    font-size: 24px;
-    font-style: normal;
-    font-weight: 600;
-    line-height: normal;
-  }
-  p {
-    color: var(--Alert-Color-Negative-Red, #ed6863);
-    /* Paragraph */
-    font-family: Pretendard;
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 20px; /* 142.857% */
-  }
-`;
-
-const Form = styled.form`
-  width: 490px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-self: stretch;
-  gap: 32px;
-`;
-
-const ButtonList = styled.div`
-  padding: 32px 0;
-  display: flex;
-  justify-content: flex-end;
-  gap: 12px;
-`;
-
-const Submit = styled.input`
-  padding: 8px 24px;
-  height: 42px;
-  background: var(--Alert-Color-Negative-Red, #ed6863);
-  border: none;
-  border-radius: 8px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-
-  color: var(--Black-White-White, #fff);
-  /* Heading 5 */
-  font-family: Pretendard;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 17px; /* 121.429% */
-`;
-
-const Cancel = styled.button`
-  padding: 8px 24px;
-  height: 42px;
-  background: var(--Black-White-White, #fff);
-  border: none;
-  border-radius: 8px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-
-  color: var(--Black-White-Black-70, #636363);
-  /* Heading 5 */
-  font-family: Pretendard;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 17px; /* 121.429% */
-`;
-
-interface DeleteProjectModalProps {
-  projectId: number;
-}
-
-const DeleteProjectModal = ({ projectId }: DeleteProjectModalProps) => {
-  const [retypeProjectTitle, setRetypeProjectTitle] = useState('');
-  const { projectData } = useGetProject(projectId);
-  const { deleteProjectMutate } = useDeleteProject();
-
-  const handleDeleteProject = async (e: React.MouseEvent<HTMLInputElement>) => {
-    e.preventDefault();
-    if (retypeProjectTitle === projectData?.title) {
-      deleteProjectMutate(projectId);
-    }
-  };
-
-  return (
-    <>
-      <ModalHeader>
-        <h2>프로젝트 삭제</h2>
-        <p>
-          프로젝트 삭제시 해당 내용을 복구할 수 없습니다. <br />
-          프로젝트 담당자를 임명하고 탈퇴 시 남은 인원이 해당 데이터를 유지할 수
-          있습니다.
-        </p>
-      </ModalHeader>
-
-      <Form>
-        <InputArea
-          value={projectData?.title || ''}
-          isDisabled={true}
-          labelText="프로젝트 명"
-          placeholderText="프로젝트 명"
-        />
-
-        <InputArea
-          value={retypeProjectTitle}
-          setValue={setRetypeProjectTitle}
-          labelText="프로젝트 재입력"
-          placeholderText="프로젝트 명을 그대로 입력해 주세요."
-        />
-
-        <ButtonList>
-          <Cancel>취소</Cancel>
-          <Submit
-            type="submit"
-            value="삭제하기"
-            onClick={handleDeleteProject}
-          />
-        </ButtonList>
-      </Form>
-    </>
-  );
-};
-
-export default DeleteProjectModal;
