+++ src/components/modal/index.tsx
@@ -1,5 +1,4 @@
 import DeleteAccountModal from './DeleteAccountModal';
-import DeleteProjectModal from './DeleteProjectModal';
 import EmailUpdateModal from './EmailUpdateModal';
 import PasswordUpdateModal from './PasswordUpdateModal';
 import ProfileUpdateModel from './ProfileUpdateModal';
@@ -7,7 +6,6 @@ import WithdrawProjectModal from './WithdrawProjectModal';
 
 export {
   DeleteAccountModal,
-  DeleteProjectModal,
   EmailUpdateModal,
   PasswordUpdateModal,
   ProfileUpdateModel,
